zmiana spriteMask na 32 bity na wiersz oraz podmiana grafiki kuli

+++ Core/gameEngine/include/sprite.h
@@ -15,7 +15,7 @@ typedef enum{
 typedef struct {
     const uint16_t *data;
     Size size;
-    const uint16_t *mask;
+    const uint32_t *mask;
 }sprite_t;
 
 /*
